consolidated development and production loggers into one file
