+++ src/logger/index.js
@@ -1,4 +1,4 @@
-const logger = require('./main')
+const logger = require('./logger')
 
 logger.info('text info', { meta: 1 })
 logger.warn('text warn')
