--- src/logger/main.js
@@ -1,13 +1,0 @@
-require('dotenv').config()
-const buildDevLogger = require('./dev-logger')
-const buildProdLogger = require('./prod-logger')
-
-let logger = null
-
-if (process.env.NODE_ENV === 'development') {
-  logger = buildDevLogger()
-} else {
-  logger = buildProdLogger()
-}
-
-module.exports = logger
